--- go/host/rpc/enclaveclient/enclave_client_user.go
@@ -1,159 +1,0 @@
-package enclaveclient
-
-import (
-	"bytes"
-	"context"
-	"errors"
-	"fmt"
-
-	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/ethereum/go-ethereum/rlp"
-	"github.com/obscuronet/go-obscuro/go/common"
-	"github.com/obscuronet/go-obscuro/go/common/rpc"
-	"github.com/obscuronet/go-obscuro/go/common/rpc/generated"
-	"github.com/obscuronet/go-obscuro/go/config"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/connectivity"
-
-	gethlog "github.com/ethereum/go-ethereum/log"
-)
-
-// EnclaveUserClient implements the common.EnclaveSystemMethods interface for internal requests (host)
-type EnclaveUserClient struct {
-	protoClient generated.EnclaveProtoClient
-	connection  *grpc.ClientConn
-	config      *config.HostConfig
-	logger      gethlog.Logger
-}
-
-func NewEnclaveUserClient(
-	protoClient generated.EnclaveProtoClient,
-	connection *grpc.ClientConn,
-	config *config.HostConfig,
-	logger gethlog.Logger,
-) common.EnclaveSystemMethods {
-	return &EnclaveUserClient{
-		protoClient: protoClient,
-		connection:  connection,
-		config:      config,
-		logger:      logger,
-	}
-}
-
-func (c *EnclaveUserClient) StopClient() error {
-	return c.connection.Close()
-}
-
-func (c *EnclaveUserClient) Status() (common.Status, error) {
-	if c.connection.GetState() != connectivity.Ready {
-		return common.Unavailable, errors.New("RPC connection is not ready")
-	}
-
-	timeoutCtx, cancel := context.WithTimeout(context.Background(), c.config.EnclaveRPCTimeout)
-	defer cancel()
-
-	resp, err := c.protoClient.Status(timeoutCtx, &generated.StatusRequest{})
-	if err != nil {
-		return common.Unavailable, err
-	}
-	if resp.GetError() != "" {
-		return common.Unavailable, errors.New(resp.GetError())
-	}
-	return common.Status(resp.GetStatus()), nil
-}
-
-func (c *EnclaveUserClient) Attestation() (*common.AttestationReport, error) {
-	timeoutCtx, cancel := context.WithTimeout(context.Background(), c.config.EnclaveRPCTimeout)
-	defer cancel()
-
-	response, err := c.protoClient.Attestation(timeoutCtx, &generated.AttestationRequest{})
-	if err != nil {
-		return nil, fmt.Errorf("failed to retrieve attestation. Cause: %w", err)
-	}
-	return rpc.FromAttestationReportMsg(response.AttestationReportMsg), nil
-}
-
-func (c *EnclaveUserClient) GenerateSecret() (common.EncryptedSharedEnclaveSecret, error) {
-	timeoutCtx, cancel := context.WithTimeout(context.Background(), c.config.EnclaveRPCTimeout)
-	defer cancel()
-
-	response, err := c.protoClient.GenerateSecret(timeoutCtx, &generated.GenerateSecretRequest{})
-	if err != nil {
-		return nil, fmt.Errorf("failed to generate secret. Cause: %w", err)
-	}
-	return response.EncryptedSharedEnclaveSecret, nil
-}
-
-func (c *EnclaveUserClient) InitEnclave(secret common.EncryptedSharedEnclaveSecret) error {
-	timeoutCtx, cancel := context.WithTimeout(context.Background(), c.config.EnclaveRPCTimeout)
-	defer cancel()
-
-	resp, err := c.protoClient.InitEnclave(timeoutCtx, &generated.InitEnclaveRequest{EncryptedSharedEnclaveSecret: secret})
-	if err != nil {
-		return err
-	}
-	if resp.GetError() != "" {
-		return errors.New(resp.GetError())
-	}
-	return nil
-}
-
-func (c *EnclaveUserClient) SubmitL1Block(block types.Block, receipts types.Receipts, isLatest bool) (*common.BlockSubmissionResponse, error) {
-	timeoutCtx, cancel := context.WithTimeout(context.Background(), c.config.EnclaveRPCTimeout)
-	defer cancel()
-
-	var buffer bytes.Buffer
-	if err := block.EncodeRLP(&buffer); err != nil {
-		return nil, fmt.Errorf("could not encode block. Cause: %w", err)
-	}
-
-	serialized, err := rlp.EncodeToBytes(receipts)
-	if err != nil {
-		return nil, fmt.Errorf("could not encode receipts. Cause: %w", err)
-	}
-
-	response, err := c.protoClient.SubmitL1Block(timeoutCtx, &generated.SubmitBlockRequest{EncodedBlock: buffer.Bytes(), EncodedReceipts: serialized, IsLatest: isLatest})
-	if err != nil {
-		return nil, fmt.Errorf("could not submit block. Cause: %w", err)
-	}
-
-	blockSubmissionResponse, err := rpc.FromBlockSubmissionResponseMsg(response.BlockSubmissionResponse)
-	if err != nil {
-		return nil, err
-	}
-	return blockSubmissionResponse, nil
-}
-
-func (c *EnclaveUserClient) SubmitBatch(batch *common.ExtBatch) error {
-	timeoutCtx, cancel := context.WithTimeout(context.Background(), c.config.EnclaveRPCTimeout)
-	defer cancel()
-
-	batchMsg := rpc.ToExtBatchMsg(batch)
-	_, err := c.protoClient.SubmitBatch(timeoutCtx, &generated.SubmitBatchRequest{Batch: &batchMsg})
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (c *EnclaveUserClient) Stop() error {
-	timeoutCtx, cancel := context.WithTimeout(context.Background(), c.config.EnclaveRPCTimeout)
-	defer cancel()
-
-	_, err := c.protoClient.Stop(timeoutCtx, &generated.StopRequest{})
-	if err != nil {
-		return fmt.Errorf("could not stop enclave: %w", err)
-	}
-	return nil
-}
-
-func (c *EnclaveUserClient) GenerateRollup() (*common.ExtRollup, error) {
-	timeoutCtx, cancel := context.WithTimeout(context.Background(), c.config.EnclaveRPCTimeout)
-	defer cancel()
-
-	resp, err := c.protoClient.CreateRollup(timeoutCtx, &generated.CreateRollupRequest{})
-	if err != nil {
-		return nil, err
-	}
-	return rpc.FromExtRollupMsg(resp.Msg), nil
-}
